feat: split load config test
